DetailView에서 내 글이 아닌 글 지우려고 하면 목록으로 제껴버리도록 기능 추가.

+++ src/main/java/board/controller/BoardDeleteController.java
@@ -3,7 +3,9 @@ package board.controller;
 import java.io.IOException;
 import java.io.PrintWriter;
 
+import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -16,6 +18,7 @@ import board.model.BoardBean;
 import board.model.BoardDao;
 import board.model.BoardReplyBean;
 import board.model.BoardReplyDao;
+import user.model.UserVo;
 
 @Controller
 public class BoardDeleteController {
@@ -44,20 +47,41 @@ public class BoardDeleteController {
 	
 	@RequestMapping(value = "/boardDelete.bd")
 	public ModelAndView doAction2(@RequestParam(value = "num") int num,
+			@RequestParam(value = "wid") String wid,
 			@RequestParam(value = "category") String category,
 			@RequestParam(value = "pageNumber") int pageNumber,
+			HttpSession session,
+			HttpServletResponse response,
+			HttpServletRequest request,
 			ModelAndView mav) throws IOException {		
 		
-			System.out.println("DELDELDEL:" + num);
-			System.out.println("DELDELDEL:" + category);
-			System.out.println("DELDELDEL:" + pageNumber);
+
+		if(session.getAttribute("loginInfo")==null) {
+			session.setAttribute("destination", "/boardInsert.bd"); // session설정해두면 아무데서나 쓸 수 있다. 목적지 설정해둔것.			
+			mav.setViewName("redirect:/login.ur"); // login.ur
+			return mav;
+		}//login null
+		else {
+			UserVo uVo =  (UserVo)session.getAttribute("loginInfo");
+			if(wid.equals(uVo.getId())){
+				System.out.println("니꺼야");
+				brDao.deleteBoardReplyData(num);			
+				bDao.deleteBoardData(num);
+				mav.setViewName("redirect:/"+category+"BoardList.bd?pageNumber="+pageNumber);
+				return mav;
+			}else {
+				System.out.println("니꺼아니야");
+				mav.setViewName("redirect:/"+category+"BoardList.bd?pageNumber="+pageNumber);
+				return mav;
+			}
 			
-			brDao.deleteBoardReplyData(num);			
-			bDao.deleteBoardData(num);
+		}
+			
+			
 
 			
-			mav.setViewName("redirect:/"+category+"BoardList.bd?pageNumber="+pageNumber);	
+				
 			
-		return mav;
+		
 	}
 }
